Initial attempt to split script into modules - messy and doesn't work just yet
Lots of import errors
